+++ cli/cli/install.sh
@@ -1,4 +1,5 @@
 #!/bin/bash
 
 # ATM, this only works on Mac because path strings are annoying across OSs.
-bash ../../set-packages.sh "" "BeamableNugetSource" "./cli/cli" "Global" ""
+cd ../..
+bash ./set-packages.sh "" "BeamableNugetSource" "./cli/cli" "Global" ""