+++ client/Packages/com.beamable/Editor/BeamCli/Commands/BeamServeCliCommandOutput.cs
@@ -1,13 +1,13 @@
 
 namespace Beamable.Editor.BeamCli.Commands
 {
-    using Beamable.Common;
-    using Beamable.Common.BeamCli;
-    
-    [System.SerializableAttribute()]
-    public class BeamServeCliCommandOutput
-    {
-        public int port;
-        public string uri;
-    }
+	using Beamable.Common;
+	using Beamable.Common.BeamCli;
+
+	[System.SerializableAttribute()]
+	public class BeamServeCliCommandOutput
+	{
+		public int port;
+		public string uri;
+	}
 }